refactor: dedicated config package
