+++ internal/config/config.go
@@ -1,4 +1,4 @@
-package api
+package config
 
 import (
 	"errors"
